1.Batch Test 오류 수정
