--- src/main/java/com/exam/batch/batch/entity/Pay2.java
@@ -1,39 +1,0 @@
-package com.exam.batch.batch.entity;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-@ToString
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-public class Pay2 {
-    private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd hh:mm:ss");
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private Long amount;
-    private String txName;
-    private LocalDateTime txDateTime;
-
-    public Pay2(Long amount, String txName, LocalDateTime txDateTime) {
-        this.amount = amount;
-        this.txName = txName;
-        this.txDateTime = txDateTime;
-    }
-
-    public static Pay2 of(Pay pay) {
-        return new Pay2(pay.getAmount(), pay.getTxName(), pay.getTxDateTime());
-    }
-}
