--- src/main/java/com/exam/batch/batch/job/reader/JpaPagingItemReaderJobConfiguration.java
@@ -1,62 +1,0 @@
-package com.exam.batch.batch.job.reader;
-
-import com.exam.batch.batch.entity.Pay;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
-import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
-import org.springframework.batch.item.ItemWriter;
-import org.springframework.batch.item.database.JpaPagingItemReader;
-import org.springframework.batch.item.database.builder.JpaPagingItemReaderBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.persistence.EntityManagerFactory;
-
-@Slf4j
-@RequiredArgsConstructor
-@Configuration
-public class JpaPagingItemReaderJobConfiguration {
-
-    private final JobBuilderFactory jobBuilderFactory;
-    private final StepBuilderFactory stepBuilderFactory;
-    private final EntityManagerFactory entityManagerFactory;
-
-    private static final int CHUNK_SIZE = 10;
-
-    @Bean
-    public Job jpaPagingItemReaderJob() {
-        return jobBuilderFactory.get("jpaPagingItemReaderJob")
-                .start(jpaPagingItemReaderStep())
-                .build();
-    }
-
-    @Bean
-    public Step jpaPagingItemReaderStep() {
-        return stepBuilderFactory.get("jpaPagingItemReaderStep")
-                .<Pay, Pay>chunk(CHUNK_SIZE)
-                .reader(jpaPagingItemReaderReader())
-                .writer(jpaPagingItemWriter())
-                .build();
-    }
-
-    @Bean
-    public JpaPagingItemReader<Pay> jpaPagingItemReaderReader() {
-        return new JpaPagingItemReaderBuilder<Pay>()
-                .name("JpaPagingItemReader")
-                .entityManagerFactory(entityManagerFactory)
-                .pageSize(CHUNK_SIZE)
-                .queryString("SELECT p FROM Pay p WHERE amount >= 2000")
-                .build();
-    }
-
-    @Bean
-    public ItemWriter<Pay> jpaPagingItemWriter() {
-        return list -> {
-            list.stream()
-                    .forEach(pay -> log.info("Current Pay={}", pay));
-        };
-    }
-}
