--- src/main/java/com/exam/batch/batch/job/step/StepNextJobConfiguration.java
@@ -1,58 +1,0 @@
-package com.exam.batch.batch.job.step;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
-import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
-import org.springframework.batch.repeat.RepeatStatus;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Slf4j
-@Configuration
-@RequiredArgsConstructor
-public class StepNextJobConfiguration {
-    public final JobBuilderFactory jobBuilderFactory;
-    public final StepBuilderFactory stepBuilderFactory;
-
-    @Bean
-    public Job stepNextJob() {
-        return jobBuilderFactory.get("stepNextJob")
-                .start(step1())
-                .next(step2())
-                .next(step3())
-                .build();
-    }
-
-    @Bean
-    public Step step1() {
-        return stepBuilderFactory.get("step1")
-                .tasklet((contribution, chunkContext) -> {
-                    log.info(">>>> This is Step1");
-                    return RepeatStatus.FINISHED;
-                })
-                .build();
-    }
-
-    @Bean
-    public Step step2() {
-        return stepBuilderFactory.get("step2")
-                .tasklet((contribution, chunkContext) -> {
-                    log.info(">>>> This is Step2");
-                    return RepeatStatus.FINISHED;
-                })
-                .build();
-    }
-
-    @Bean
-    public Step step3() {
-        return stepBuilderFactory.get("step3")
-                .tasklet((contribution, chunkContext) -> {
-                    log.info(">>>> This is Step3");
-                    return RepeatStatus.FINISHED;
-                })
-                .build();
-    }
-}
