--- src/main/java/com/exam/batch/batch/job/writer/JdbcBatchItemWriterJobConfiguration.java
@@ -1,69 +1,0 @@
-package com.exam.batch.batch.job.writer;
-
-import com.exam.batch.batch.entity.Pay;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
-import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
-import org.springframework.batch.item.database.JdbcBatchItemWriter;
-import org.springframework.batch.item.database.JdbcCursorItemReader;
-import org.springframework.batch.item.database.builder.JdbcBatchItemWriterBuilder;
-import org.springframework.batch.item.database.builder.JdbcCursorItemReaderBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
-
-import javax.sql.DataSource;
-
-@Slf4j
-@RequiredArgsConstructor
-@Configuration
-public class JdbcBatchItemWriterJobConfiguration {
-
-    private final JobBuilderFactory jobBuilderFactory;
-    private final StepBuilderFactory stepBuilderFactory;
-    private final DataSource dataSource;
-    private static final int CHUNK_SIZE = 10;
-
-    @Bean
-    public Job jdbcBatchItemWriterJob() {
-        return jobBuilderFactory.get("jdbcBatchItemWriterJob")
-                .start(jdbcBatchItemWriterStep())
-                .build();
-    }
-
-    @Bean
-    public Step jdbcBatchItemWriterStep() {
-        return stepBuilderFactory.get("jdbcBatchItemWriterStep")
-                .<Pay, Pay>chunk(CHUNK_SIZE)
-                .reader(jdbcBatchItemWriterReader())
-                .writer(jdbcBatchItemWriter())
-                .build();
-    }
-
-    @Bean
-    public JdbcCursorItemReader<Pay> jdbcBatchItemWriterReader() {
-        return new JdbcCursorItemReaderBuilder<Pay>()
-                .fetchSize(CHUNK_SIZE)
-                .dataSource(dataSource)
-                .rowMapper(new BeanPropertyRowMapper<>(Pay.class))
-                .sql("SELECT id, amount, tx_name, tx_date_time FROM pay")
-                .name("jdbcBatchItemWriter")
-                .build();
-    }
-
-    /**
-     * Reader에서 넘어온 데이터를 하나씩 출력한다
-     * JdbcBatchItemWriter<Pay>에서 Pay는 Reader에서 넘겨주는 값의 타입이다.
-     */
-    @Bean // beanMapped()를 사용시 필수
-    public JdbcBatchItemWriter<Pay> jdbcBatchItemWriter() {
-        return new JdbcBatchItemWriterBuilder<Pay>()
-                .dataSource(dataSource)
-                .sql("insert into pay2(amount, tx_name, tx_date_time) values (:amount, :txName, :txDateTime)") // 맵핑되는 필드명은 (:amount등 ) dto의 getter와 매핑됨
-                .beanMapped() // insert SQL value를 Pay객체와 맵핑함
-                .build();
-    }
-}
