--- src/main/java/com/exam/batch/batch/job/writer/JpaItemWriterJobConfiguration.java
@@ -1,70 +1,0 @@
-package com.exam.batch.batch.job.writer;
-
-import com.exam.batch.batch.entity.Pay;
-import com.exam.batch.batch.entity.Pay2;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
-import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
-import org.springframework.batch.item.ItemProcessor;
-import org.springframework.batch.item.ItemWriter;
-import org.springframework.batch.item.database.JpaItemWriter;
-import org.springframework.batch.item.database.JpaPagingItemReader;
-import org.springframework.batch.item.database.builder.JpaPagingItemReaderBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.persistence.EntityManagerFactory;
-
-@Slf4j
-@RequiredArgsConstructor
-@Configuration
-public class JpaItemWriterJobConfiguration {
-
-    private final JobBuilderFactory jobBuilderFactory;
-    private final StepBuilderFactory stepBuilderFactory;
-    private final EntityManagerFactory entityManagerFactory;
-
-    private static final int CHUNK_SIZE = 10;
-
-    @Bean
-    public Job jpaItemWriterJob() {
-        return jobBuilderFactory.get("jpaItemWriterJob")
-                .start(jpaItemWriterStep())
-                .build();
-    }
-
-    @Bean
-    public Step jpaItemWriterStep() {
-        return stepBuilderFactory.get("jpaItemWriterStep")
-                .<Pay, Pay2>chunk(CHUNK_SIZE)
-                .reader(jpaItemWriterReader())
-                .processor(jpaItemProcessor())
-                .writer(jpaItemWriter())
-                .build();
-    }
-
-    @Bean
-    public JpaPagingItemReader<Pay> jpaItemWriterReader() {
-        return new JpaPagingItemReaderBuilder<Pay>()
-                .name("jpaItemWriterReader")
-                .entityManagerFactory(entityManagerFactory)
-                .pageSize(CHUNK_SIZE)
-                .queryString("SELECT p FROM Pay p")
-                .build();
-    }
-
-    @Bean
-    public ItemProcessor<Pay, Pay2> jpaItemProcessor() {
-        return pay -> new Pay2(pay.getAmount(), pay.getTxName(), pay.getTxDateTime());
-    }
-
-    @Bean
-    public ItemWriter<Pay2> jpaItemWriter() {
-        JpaItemWriter<Pay2> jpaItemWriter = new JpaItemWriter<>();
-        jpaItemWriter.setEntityManagerFactory(entityManagerFactory);
-        return jpaItemWriter;
-    }
-}
