+++ src/main/java/com/exam/batch/job/decider/DeciderJobConfiguration.java
@@ -1,0 +1,73 @@
+package com.exam.batch.job.decider;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
+import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
+import org.springframework.batch.core.job.flow.JobExecutionDecider;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * decider 결과에 따라 지정된 step을 실행하는 배치 클래스
+ */
+@Slf4j
+@Configuration
+@RequiredArgsConstructor
+public class DeciderJobConfiguration {
+    public final JobBuilderFactory jobBuilderFactory;
+    public final StepBuilderFactory stepBuilderFactory;
+
+    @Bean
+    public Job decideJob() {
+        return jobBuilderFactory.get("decideJob")
+                .start(startStep())// 스탭실행
+                .next(decider())// 실행한 결과가 홀수인지 짝수인지 판단
+                .from(decider())/// 홀수면 홀수 스탭진행
+                    .on("ODD")
+                    .to(oddStep())
+                .from(decider()) // 짝수면 짝수스탭 진행
+                    .on("EVEN")
+                    .to(evenStep())
+                .end()
+                .build();
+    }
+
+    @Bean
+    public Step startStep() {
+        return stepBuilderFactory.get("startStep")
+                .tasklet((contribution, chunkContext) -> {
+                    log.info(">>> Start");
+                    return RepeatStatus.FINISHED;
+                })
+                .build();
+    }
+
+    @Bean
+    public JobExecutionDecider decider() {
+        return new OddDecider();
+    }
+
+    @Bean
+    public Step oddStep() {
+        return stepBuilderFactory.get("oddStep")
+                .tasklet((contribution, chunkContext) -> {
+                    log.info(">>>oddStep Start");
+                    return RepeatStatus.FINISHED;
+                })
+                .build();
+    }
+
+    @Bean
+    public Step evenStep() {
+        return stepBuilderFactory.get("evenStep")
+                .tasklet((contribution, chunkContext) -> {
+                    log.info(">>>evenStep Start");
+                    return RepeatStatus.FINISHED;
+                })
+                .build();
+    }
+}
