+++ src/main/java/com/exam/batch/job/processor/ProcessorCompositeJobConfiguration.java
@@ -1,0 +1,97 @@
+package com.exam.batch.job.processor;
+
+import com.exam.batch.entity.Pay;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
+import org.springframework.batch.core.configuration.annotation.JobScope;
+import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
+import org.springframework.batch.item.ItemProcessor;
+import org.springframework.batch.item.ItemReader;
+import org.springframework.batch.item.ItemWriter;
+import org.springframework.batch.item.database.builder.JpaPagingItemReaderBuilder;
+import org.springframework.batch.item.support.CompositeItemProcessor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.persistence.EntityManagerFactory;
+import java.util.ArrayList;
+import java.util.List;
+
+@Slf4j
+@RequiredArgsConstructor
+@Configuration
+public class ProcessorCompositeJobConfiguration {
+
+    public static final String JOB_NAME = "processorCompositeBatch";
+    public static final String BEAN_PREFIX = JOB_NAME + "_";
+
+    private final JobBuilderFactory jobBuilderFactory;
+    private final StepBuilderFactory stepBuilderFactory;
+    private final EntityManagerFactory entityManagerFactory;
+
+    @Value("${chunkSize:1000}")
+    private int CHUNK_SIZE;
+
+    @Bean(JOB_NAME)
+    public Job job() {
+        return jobBuilderFactory.get(JOB_NAME)
+                .preventRestart()
+                .start(step())
+                .build();
+    }
+
+    @Bean(BEAN_PREFIX + "step")
+    @JobScope
+    public Step step() {
+        return stepBuilderFactory.get(BEAN_PREFIX + "step")
+                .<Pay, String>chunk(CHUNK_SIZE)
+                .reader(reader())
+                .processor(compositeItemProcessor())
+                .writer(writer())
+                .build();
+    }
+
+    @Bean(BEAN_PREFIX + "reader")
+    public ItemReader<Pay> reader() {
+        return new JpaPagingItemReaderBuilder<Pay>()
+                .name(BEAN_PREFIX + "reader")
+                .entityManagerFactory(entityManagerFactory)
+                .pageSize(CHUNK_SIZE)
+                .queryString("SELECT p FROM Pay p")
+                .build();
+    }
+
+    @Bean
+    public CompositeItemProcessor compositeItemProcessor() {
+        List<ItemProcessor> delegates = new ArrayList<>(2);
+        delegates.add(processor1());
+        delegates.add(processor2());
+
+        CompositeItemProcessor processor = new CompositeItemProcessor<>(); // 제네릭 타입을 쓰면 processor 의 모든 타입이 같아야됨
+
+        processor.setDelegates(delegates);
+
+        return processor;
+    }
+
+    private ItemProcessor<Pay, String> processor1() {
+        log.info("프로세서1");
+        return Pay::getTxName;
+    }
+
+    private ItemProcessor<String, String> processor2() {
+        log.info("프로세서2");
+        return name -> "안녕하세요" + name + "입니다";
+    }
+
+    private ItemWriter<String> writer() {
+        return items -> {
+            items.stream()
+                    .forEach(item -> log.info("item={}", item));
+        };
+    }
+}
