+++ src/main/java/com/exam/batch/job/writer/CustomItemWriterJobConfiguration.java
@@ -1,0 +1,66 @@
+package com.exam.batch.job.writer;
+
+import com.exam.batch.entity.Pay;
+import com.exam.batch.entity.Pay2;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
+import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
+import org.springframework.batch.item.ItemProcessor;
+import org.springframework.batch.item.ItemReader;
+import org.springframework.batch.item.ItemWriter;
+import org.springframework.batch.item.database.builder.JpaPagingItemReaderBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.persistence.EntityManagerFactory;
+
+@Slf4j
+@RequiredArgsConstructor
+@Configuration
+public class CustomItemWriterJobConfiguration {
+
+    private final JobBuilderFactory jobBuilderFactory;
+    private final StepBuilderFactory stepBuilderFactory;
+    private final EntityManagerFactory entityManagerFactory;
+
+    private static final int CHUNK_SIZE = 10;
+
+    @Bean
+    public Job customItemWriterJob() {
+        return jobBuilderFactory.get("customItemWriterJob")
+                .start(customItemWriterStep())
+                .build();
+    }
+
+    private Step customItemWriterStep() {
+        return stepBuilderFactory.get("customItemWriterStep")
+                .<Pay, Pay2> chunk(CHUNK_SIZE)
+                .reader(customItemWriterReader())
+                .processor(customItemWriterProcessor())
+                .writer(customItemWriterWriter())
+                .build();
+    }
+
+    private ItemReader<Pay> customItemWriterReader() {
+        return new JpaPagingItemReaderBuilder<Pay>()
+                .name("customItemWriterReader")
+                .entityManagerFactory(entityManagerFactory)
+                .pageSize(CHUNK_SIZE)
+                .queryString("SELECT p FROM Pay p")
+                .build();
+    }
+
+    private ItemProcessor<Pay, Pay2> customItemWriterProcessor() {
+        return pay -> new Pay2(pay.getAmount(), pay.getTxName(), pay.getTxDateTime());
+    }
+
+    private ItemWriter<Pay2> customItemWriterWriter() {
+        return items -> {
+            items.stream()
+                    .forEach(item -> log.info("item={}",item));
+        };
+    }
+}
