+++ src/main/java/com/exam/batch/repository/PayRepository.java
@@ -1,0 +1,7 @@
+package com.exam.batch.repository;
+
+import com.exam.batch.entity.Pay;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PayRepository extends JpaRepository<Pay, Long> {
+}
