--- src/test/java/com/exam/batch/batch/BatchApplicationTests.java
@@ -1,13 +1,0 @@
-package com.exam.batch.batch;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BatchApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
